+++ src/components/EditForm/index.jsx
@@ -1,0 +1,1 @@
+export { EditForm } from "src/components/EditForm/EditForm";
