+++ src/components/ItemWithEditForm/index.jsx
@@ -1,0 +1,1 @@
+export { ItemWithEditForm } from "src/components/ItemWithEditForm/ItemWithEditForm";
